static tasks

+++ app/src/main/java/com/platepicks/util/ImageSaver.java
@@ -26,11 +26,11 @@ public class ImageSaver {
 
     private String directoryName = "images";
     private String[] fileNames;
-    private WeakReference<Context> contextRef;
+    private Context context;
     boolean small = false;
 
     public ImageSaver(Context context) {
-        this.contextRef = new WeakReference<>(context);
+        this.context = context;
     }
 
     public ImageSaver setFileName(String... fileNames) {
@@ -44,25 +44,25 @@ public class ImageSaver {
     }
 
     public void save(Bitmap... bitmapImage) {
-        new SaveImageTask().execute(bitmapImage);
+        new SaveImageTask(context, directoryName, fileNames).execute(bitmapImage);
     }
 
-    private File createFile(String file) {
-        if (contextRef == null || contextRef.get() == null)
+    private static File createFile(Context context, String directoryName, String file) {
+        if (context == null || directoryName == null || file == null)
             return null;
 
-        File directory = contextRef.get().getDir(directoryName, Context.MODE_PRIVATE);
+        File directory = context.getDir(directoryName, Context.MODE_PRIVATE);
 //        for (String s : directory.list())
 //            Log.d("ImageSaver", "File: " + s);
         return new File(directory, file);
     }
 
     public void load(ImageView imageView, OnCompleteListener caller, boolean small) {
-        Log.d("ImageSaver", " In Load");
+        Log.d("ImageSaver", fileNames[0] + " In Load");
         this.small = small;
 
-        if (cancelPotentialLoad(imageView)) {
-            LoadImageAsyncTask loadTask = new LoadImageAsyncTask(imageView, caller, fileNames[0]);
+        if (cancelPotentialLoad(imageView, fileNames[0])) {
+            LoadImageAsyncTask loadTask = new LoadImageAsyncTask(imageView, caller, context, directoryName, fileNames[0], small);
             DownloadedDrawable drawable = new DownloadedDrawable(loadTask);
             imageView.setImageDrawable(drawable);
             drawable.getLoadImageAsyncTask().execute();
@@ -70,15 +70,15 @@ public class ImageSaver {
     }
 
     public void delete() {
-        new DeleteImageAsyncTask().execute();
+        new DeleteImageAsyncTask(context, directoryName, fileNames[0]).execute();
     }
 
-    private boolean cancelPotentialLoad(ImageView imageView) {
+    private static boolean cancelPotentialLoad(ImageView imageView, String filename) {
         LoadImageAsyncTask loadTask = getLoadImageAsyncTask(imageView);
 
         if (loadTask != null) {
             String bitmapFile = loadTask.filename;
-            if (!bitmapFile.equals(fileNames[0])) {
+            if (!bitmapFile.equals(filename)) {
                 loadTask.cancel(true);
                 Log.d("ImageSaver", "Cancelled previous task");
             } else {
@@ -90,7 +90,7 @@ public class ImageSaver {
         return true;
     }
 
-    private LoadImageAsyncTask getLoadImageAsyncTask(ImageView imageView) {
+    private static LoadImageAsyncTask getLoadImageAsyncTask(ImageView imageView) {
         if (imageView != null) {
             Drawable drawable = imageView.getDrawable();
             if (drawable instanceof DownloadedDrawable) {
@@ -101,7 +101,17 @@ public class ImageSaver {
         return null;
     }
 
-    class SaveImageTask extends AsyncTask<Bitmap, Void, Void> {
+    static class SaveImageTask extends AsyncTask<Bitmap, Void, Void> {
+        WeakReference<Context> contextRef;
+        String directoryName;
+        String[] filenames;
+
+        public SaveImageTask(Context context, String directoryName, String[] filenames) {
+            this.contextRef = new WeakReference<>(context);
+            this.directoryName = directoryName;
+            this.filenames = filenames;
+        }
+
         @Override
         protected Void doInBackground(Bitmap... params) {
             if (params.length < 1)
@@ -111,12 +121,14 @@ public class ImageSaver {
             try {
                 accessFiles.lock();
                 for (int i = 0; i < params.length; i++) {
-                    File newImg = createFile(fileNames[i]);
+                    if (contextRef != null && contextRef.get() != null) {
+                        File newImg = createFile(contextRef.get(), directoryName, filenames[i]);
 
-                    if (newImg != null && !newImg.exists()) {
-                        newImg.deleteOnExit();
-                        fileOutputStream = new FileOutputStream(newImg);
-                        params[i].compress(Bitmap.CompressFormat.PNG, 100, fileOutputStream);
+                        if (newImg != null && !newImg.exists()) {
+                            newImg.deleteOnExit();
+                            fileOutputStream = new FileOutputStream(newImg);
+                            params[i].compress(Bitmap.CompressFormat.PNG, 100, fileOutputStream);
+                        }
                     }
                 }
                 accessFiles.unlock();
@@ -135,16 +147,21 @@ public class ImageSaver {
         }
     }
 
-    class LoadImageAsyncTask extends AsyncTask<Void, Void, Bitmap> {
+    static class LoadImageAsyncTask extends AsyncTask<Void, Void, Bitmap> {
         WeakReference<ImageView> imageViewRef;  // To not stop garbage collection if imageview is gone
         WeakReference<OnCompleteListener> callerRef;
-
+        WeakReference<Context> contextRef;
+        String directoryName;
         String filename;
+        boolean small;
 
-        LoadImageAsyncTask(ImageView imageView, OnCompleteListener caller, String filename) {
+        LoadImageAsyncTask(ImageView imageView, OnCompleteListener caller, Context context, String directoryName, String filename, boolean small) {
             this.imageViewRef = new WeakReference<>(imageView);
             this.callerRef = new WeakReference<>(caller);
+            this.contextRef = new WeakReference<>(context);
+            this.directoryName = directoryName;
             this.filename = filename;
+            this.small = small;
         }
 
         @Override
@@ -153,8 +170,8 @@ public class ImageSaver {
 
             FileInputStream inputStream = null;
             try {
-                File imgFile = createFile(filename);
                 accessFiles.lock();
+                File imgFile = createFile(contextRef.get(), directoryName, filename);
 
                 if (imgFile == null) {
                     Log.e("ImageSaver Load", "Null returned file. Context garbage collected?");
@@ -204,13 +221,29 @@ public class ImageSaver {
         }
     }
 
-    class DeleteImageAsyncTask extends AsyncTask<Void, Void, Void> {
+    static class DeleteImageAsyncTask extends AsyncTask<Void, Void, Void> {
+        WeakReference<Context> contextRef;
+        String directoryName;
+        String filename;
+
+        public DeleteImageAsyncTask(Context context, String directoryName, String filename) {
+            this.contextRef = new WeakReference<Context>(context);
+            this.directoryName = directoryName;
+            this.filename = filename;
+        }
+
         @Override
         protected Void doInBackground(Void... params) {
             accessFiles.lock();
 
             Log.d("ImageSaver", "In delete");
-            File file = createFile(fileNames[0]);
+
+            if (contextRef == null || contextRef.get() == null) {   // Ensure not null
+                accessFiles.unlock();
+                return null;
+            }
+
+            File file = createFile(contextRef.get(), directoryName, filename);
 
             if (file == null) {
                 Log.e("ImageSaver Delete", "Null returned file. Context garbage collected?");
@@ -220,9 +253,9 @@ public class ImageSaver {
 
             if (file.exists()) {
                 if (!file.delete())
-                    Log.e("ImageSaver", "Failed to delete " + fileNames[0]);
+                    Log.e("ImageSaver", "Failed to delete " + filename);
                 else
-                    Log.d("ImageSaver", "Deleted " + fileNames[0]);
+                    Log.d("ImageSaver", "Deleted " + filename);
             }
 
             accessFiles.unlock();
